--- code/ai-service/fast-api/vercel.json
@@ -1,14 +1,0 @@
-{
-    "builds": [
-        {
-            "src": "api.py",
-            "use": "@vercel/python"
-        }
-    ],
-    "routes": [
-        {
-            "src": "/(.*)",
-            "dest": "api.py"
-        }
-    ]
-}
